ovl: flag inodes with persistent and consistent inode number
With xino feature enabled or with all layers on the same fs,
overlay tries to create a consistent and persistent inode number
for every overlayfs inode.

Currently, we flag overlayfs inodes whose inode number should survive
across copy up, but that inode number can be short lived and may only
be unique in the namespace of a specific lower fs pseudo st_dev.

Use another flag to mark overlayfs inodes that have a persistent inode
number assignment in the unified namespace of overlay st_dev.

In that case, the persistency of the inode number across copy up and
mount cycle also depends on lower filesystem support for file handles.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>