ovl: store persistent inode number in origin xattr
When lower filesystem does not support decoding file handles, we cannot
find the origin inode of a non-directory by file handle and we will lose
the persistent overlayfs inode number after a mount cycle.

In this case, we currently store a 'null' fh in origin xattr to signify
that the file or directory were copied up.

In case we have an inode number that would have been persistent if file
handles were supported, store the inode number in the origin xattr
instead of a 'null' fh, so we can later use it to preserve the inode
number across mount cycle.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>