ovl: preparing for marking a merge dir opaque
A pure upper directory is created opaque inside a merge parent and can
become opaque when moved into a merge parent.
This type of overlay directory object has no lower stack.

The only way for an overlay directory with lower stack to become opaque
is when it is empty and just before it is being removed.

Prepare for being able to make a merge directory opaque when there is
nothing important to be found in the lower stack.

Teach ovl_type_merge() to also examine the opaque dentry flag in
addition to ovl_lowernum() and use ovl_type_merge*() helpers in the
routines that iterate the merge directory lower stack.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>