ovl: set merge directory opaque when lower stack is useless
With xino=nofollow, we allow st_ino of merge directory to change to the
upper st_ino in the same manner that a non-dir changes its st_ino on
copy up when all the children of the directory are "fully copied up"
by making the directory opaque.

Note that "fully copied up" for sub-directories means that those
sub-directories are pure upper or "opaque".

The opportunity the check if a merge directory can be made opaque is
when creating the readdir cache and before a directory is made opaque
all of its sub-directory need to be made opaque, so the overlay needs to
be iterated in DFS order for all the directories to be made opaque.

After all files in an overlay have been copied up and all directories
have been made opaque, the lower layer fs can be detached using shutdown
IOCTL or network/FUSE force disconnect, because none of the lower inodes
will be accessed by overlayfs.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>