fsnotify: enable notifcations on overlayfs real files
Overlayfs opens the real underlying files with fake path - whose f_inode
in on the underlying fs and f_path on overlayfs.

Those real files are open with FMODE_NONOTIFY, because fsnotify code was
not prapared to handle fanotify hooks on files with fake path correctly
and fanotify would report unexpected event->fd with fake overlayfs path.

Teach fsnotify how to deal with real files with fake path correctly and
remove FMODE_NONOTIFY from real files.

If there is an fanotify listener on the underlying fs which expects
event->fd, the events on the real files with fake path are silently
dropped, but if the listener is handling permission events, the
operation on real file that would report a permission event is denied.

If the fanotify listener was initialized with FAN_REPORT_FID and does
not expect event->fd, the events on the real files will be reported
with file ids, just as overlayfs operations on real inodes and dentries
are reported with file ids.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>