fs: use fake_file container for internal files with fake f_path
Overlayfs and cachefiles use open_with_fake_path() to allocate internal
files, where overlayfs also puts a "fake" path in f_path - a path which
is not on the same fs as f_inode.

Allocate a container struct file_fake for those internal files, that
is used to hold the fake path along with an optional real path.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>