fanotify: implement FAN_PRE_ACCESS event
Similar to FAN_ACCESS_PERM permission event, but it is only allowed with
class FAN_CLASS_PRE_CONTENT and unlike FAN_ACCESS_PERM, it is safe to
write to the file being accessed in the context of the permission event.

This permission event is meant to be used by hierarchical storage
managers that want to fill the content of files on first access.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>