fsnotify: optimize the case of no pre-modify event watchers
Commit e43de7f0862b ("fsnotify: optimize the case of no marks of any type")
optimized the case where there are no fsnotify watchers on any of the
filesystem's objects.

It is quite common for a system to have a single local filesystem and
it is quite common for the system to have some inotify watches on some
config files or directories, so the optimization of no marks at all is
often not in effect.

fanotify pre-modify permission watchers are going to be more rare and
have more impact, so optimizing the case of no marks with pre-modify
events could improve performance for more systems.

Maintain a per-sb counter of objects that have marks with pre-modify
events in their mask and use that counter to optimize out pre-modify
permission hooks.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>