fs: create helpers {sb,file}_write_not_started()
Create new helpers {sb,file}_write_not_started() that can be used to
assert that sb_start_write() is not held.

These are not the same as the negation of {sb,file}_write_started(),
because unlike the old helpers, if unsure, the new helpers indicate that
sb_start_write() is not held.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>