fsnotify: introduce FS_PRE_ACCESS event
In new kernels, FS_PRE_ACCESS event is identical to FS_ACCESS_PERM event,
but it adds a semantic statement, that it is safe to write to the file
being accessed in the context of the permission event handler.

This permission event is meant to be used by hierarchical storage
managers that want to fill the content of files on first access.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>