fuse: implement ioctls to manage backing files
FUSE server calls the FUSE_DEV_IOC_BACKING_OPEN ioctl with a backing
file descriptor. If the call succeeds, it receives back an identifier.
This identifier is used when replying to OPEN request with the flag
FOPEN_PASSTHROUGH to setup passthrough of file operations on the
open FUSE file to the associated backing file.

The FUSE server may call FUSE_DEV_IOC_BACKING_CLOSE ioctl to unmap
the the backing file id. This can be done at any time, but if an open
reply with the backing id is still in progress, the open may or may not
end up settting up the passthrough to the backing file.

In any case, the backing file will be kept open by the FUSE driver until
the last fuse_file that was setup to passthrough to that backing file is
closed AND the FUSE_DEV_IOC_BACKING_CLOSE ioctl was called.

Setting up backing file requires a server with CAP_SYS_ADMIN privileges.
For the backing file to be successfully setup, the backing file must
implement both read_iter and write_iter file operations.

The limitation on the level of filesystem stacking allowed for the
backing file is enforced before setting up the backing file id.

Signed-off-by: Alessio Balsini <balsini@android.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>