fuse: implement ioctls to setup inode bound backing files
Introduce the FUSE_DEV_IOC_BACKING_OPEN ioctl, which allows the server
to setup a 1-to-1 association between a FUSE inode and a backing file.
These files will be referred to as "inode bound backing files".

Setting up backing file requires a server with CAP_SYS_ADMIN privileges.
If an inode bound backing file already exists, the ioctl returns -EEXIST.
Managing which thread sets up the backing file for concurrent file open
requests is the responsibility of the server.

This association is used when replying to OPEN with the flag
FOPEN_PASSTHROUGH to setup passthrough of file operations on the
open FUSE file to the associated backing file.

For the association to be successfully setup, the backing file must
implement both read_iter and write_iter file operations.
This extra check avoids special pseudo files to be used for passhrough.
The limitation on the level of filesystem stacking allowed for the
backing file is enforced before setting up the association.

The FUSE server may call FUSE_DEV_IOC_BACKING_CLOSE ioctl to break
the association between the backing file and the inode.

If there is no inode bound backing file, the ioctl returns -ENOENT.
Managing which thread detaches the backing file is the responsibility of
the server.

Even if the server doesn't call FUSE_DEV_IOC_BACKING_CLOSE ioctl, the
inode bound backing file will be released when the FUSE inode is evicted.

In any case, the backing file will be kept open by the FUSE driver until
the last FUSE file that was setup to passthrough to that backing file is
closed.

[Amir: This is a variant of Alessio's original patch which implementes
       the alternative design of refcounted inode bound backing files.]

Signed-off-by: Alessio Balsini <balsini@android.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>