fuse: implement ioctls to setup unbound backing files
Implement a mode for FUSE_DEV_IOC_BACKING_OPEN ioctl to associate
a backing file with a backing file id, whose lifetime is not bound to
a specific inode. This will be referred to as an "unbound backing file".

The unbound backing file can be used to passthrough FUSE files from
several different FUSE inodes to the same backing file, or several
FUSE files from the same FUSE inode, to different backing files.

To setup an unbound backing file the FUSE server needs to call the
FUSE_DEV_IOC_BACKING_OPEN ioctl with the FUSE_BACKING_MAP_ID flag and,
if the call succeeds, receives back an identifier that will be used at
open/create response time in the fuse_open_out backing_id field to
associate the FUSE file to the backing file.

The FUSE server may call FUSE_DEV_IOC_BACKING_CLOSE ioctl to unmap
the the backing file id. This can be done on file release or later.

In any case, the backing file will be kept open by the FUSE driver until
the last fuse_file that was setup to passthrough to that backing file is
closed AND the FUSE_DEV_IOC_BACKING_CLOSE ioctl was called.

Signed-off-by: Alessio Balsini <balsini@android.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>