fuse: implement auto close of backing files ids
Implement the flag FOPEN_CLOSE_BACKING_ID for open, as an extention to
FUSE_PASSTHROUGH, relevant for passthrough to unbound backing files.
The macro FOPEN_PASSTHROUGH_AUTO_CLOSE combines these two flags.

In addition to setting up passthtough to the unbound backing file, the
open response has the effect of the FUSE_DEV_IOC_BACKING_CLOSE ioctl.

Unlike FUSE_PASSTHROUGH, which takes a refcount on the unbound backing
file, the backing file id is unmapped and its refcount is transfered to
the FUSE file object.

The backing file itself will be closed when the last FUSE file that was
setup to passthrough to that backing file is released.

FUSE server can use this flag to simplify the management of backing files.
The server allocates a backing id for each open request, and responds with
FOPEN_PASSTHROUGH_AUTO_CLOSE, so it is not responsible of unmapping those
backing file ids.

The FOPEN_CLOSE_BACKING_ID flag is ignored if the open response does not
contain a valid backing id. Specifically, the flag is ignored when open
requests passthrough to inode bound backing files (i.e. backing id 0).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>