fuse: implement auto close of backing files ids
Implement the flag FOPEN_CLOSE_BACKING_ID for open, as an extention to
FUSE_PASSTHROUGH - the macro FOPEN_PASSTHROUGH_AUTO_CLOSE combines these
two flags.

In addition to setting up passthtough to the backing file, the open
response has the effect of the FUSE_DEV_IOC_BACKING_CLOSE ioctl.

Unlike FUSE_PASSTHROUGH, which takes a refcount on the backing file,
the backing file id is unmapped and its refcount is transfered to the
FUSE file object.

The backing file itself will be closed when the last FUSE file that was
setup to passthrough to that backing file is released.

FUSE server can use this flag to simplify the management of backing files.
The server allocates a backing id for each open request, and responds with
FOPEN_PASSTHROUGH_AUTO_CLOSE, so it is not responsible of unmapping those
backing file ids.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>