ecryptfs: unshare common stackable filesystem helpers
eCryptfs uses some helpers for stackable filesystems in fs/stack.c.

The introduction commit for fs/stack.c mentions that the helpers are for
"stackable filesystems (such as eCryptfs and Unionfs)", but there is no
Unionfs and eCryptfs was the only user of those helpers.

Furthermore, the helpers have logic that is specific to eCryptfs, so
overlayfs cannot use them.

Unshare the eCryptfs helpers and remove fs/stack.c, so we can start fresh
with another attempt to create a useful library for stackable filesystems.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>