fuse: implement read/write passthrough
Use the backing file read/write helpers to implement read/write
passthrough to a backing file.

Similar to update size/mtime at the end of fuse_perform_write(),
we need to bump the attr version when we update the inode size,
so extend the ->end_write() callback to report pos and number of
bytes written.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>