fuse: prepare for long lived reference on backing file
Currently backing file is attached to fuse inode on the first
passthrough open of the inode and detached on last passthourgh close.

In preparation for attaching a backing file on lookup, allow attaching
a single long lived reference on fuse inode backing file that will be
used for passthrough of inode operations and dropped on fuse inode evict.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>