fuse: prepare for passthrough of inode operations
So far, fuse passthrough was implemented for read/write/splice/mmap
operations for files opened with FOPEN_PASSTHROUGH.

A backing file is attached to a fuse inode, but only for as long as
there are FOPEN_PASSTHROUGH files opens.

We would like to attach a backing file to fuse inode on lookup and
passthrough some inode operations.

Add field ops_mask to the input argument of FUSE_DEV_IOC_BACKING_OPEN
ioctl to declare the operations that would passthrough to the backing
file once it has been attached to the fuse inode.

Setting the FUSE_READ/FUSE_WRITE operations in the ops_mask is not
required because those operations are implied by FOPEN_PASSTHROUGH.

Relax the sanity check for ->{read,write}_iter() file ops in
FUSE_DEV_IOC_BACKING_OPEN ioctl to allow using an O_PATH fd
for passthrough of inode operations.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>