fuse: take long lived reference to backing file for inode ops passthrough
When a backing file is setup for inode operations passthrough, take a
long lived refcount when backing file is attached to inode on first
passthrough open, so that the backing file could be used for inode ops
even after the file is closed.

This final refount will be dropped on inode evict.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>