ovl: only revalidate the uppermost real dentry in the stack
Overlayfs does not support modifications to underlying layers and the
behavior in this case is undefined.

When layers are on a network fs (or fuse), changes on the server also
qualify as underlying layer modification, which are not officially
supported, and yet overlayfs has always proxied the ->d_revalite()
call to all underlying layers, so it is not unreasonable that users have
been depending on changes on server in overalyfs layers being revalidated.

We would like to introduce a change to the ->d_revalidate() API, which
passes the expected parent+name explicitly.

The overlayfs dentry name is always the same as the uppermost real dentry
name, so revalidating the uppermost real dentry is easy.

Revalidating all the dentries in the stack is more challenging and
brings less value.

Therefore, as best effort to preserve legacy behavior, at least for
LOOKUP_CREATE | LOOKUP_RENAME_TARGET, call d_revalidate() only on the
uppermost real dentry, which determines the overlay dentry name and
do not revalidating all the dentries in the stack.

When uppermost real dentry is a hardlink (or an indexed copy up), the
overlay dentry may have a different name than the real dentry, so skip
the best effort revalidate in this case.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>