ovl: "finalize" upper dir when all child entries are copied up
When merged readdir detects that all merged dir children are copied up,
we "finalize" the upeer dir content, so we won't need to iterate lower
dirs on next readdir.

If any of the new overlayfs features are enabled (e.g. metacopy), we
do not support offline modification of lower layers, so we mark the
upper dir permanently "finalized" with the special xattr value
overlay.opaque="z".  This overloading of overlay.opaque xattr is fine
because a "finalized" upper dir can never be opaque (y) and upper dir
cannot have xwhiteouts (x).

This format means that middle layers can be composed offline with either
xwhiteouts or with "finalized" directories, not with both.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>