ovl: remove snapshot index after delete of lowerdir
This is best effort - we can cope with leftover index entries if we
fail to remove index emtry or if we crash before removing index entry.

When watching mount, we must add a mark on sb for FS_DELETE event,
because FS_DELETE event cannot be delivered to mount mark and then
we also need to add an ignore mark on the index dir to suppress the
delete events from deleted index entries.

Change hook to deliver delete event before unlinking dentry for the
debug prints and for filtering by path.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>