ovl: relax assertions on failure to encode file handles
Encoding file handles is usually performed by a filesystem ->encode_fh()
method that may fail for various reasons.

The legacy callers of exportfs_encode_fh(), namely, nfsd and
name_to_handle_at(2) syscall are ready to cope with the possibility
of failure to encode a file handle.

Unlike those legacy callers, ovl_encode_real_fh() has a WARN_ON()
assertion when ->encode_fh() fails.

Relax this assertion because it is wrong.

The syzbot bug report is probably a result of commit 16aac5ad1fa9
("ovl: support encoding non-decodable file handles") increasing the
chances of the assertion.

Triggering this assertion was always possible with other filesystems and
other reasons of ->encode_fh() failures and more particularly, it was
also possible with the exact same reproducer using overlayfs that is
mounted with options index=on,nfs_export=on also on kernels < v6.6.
Therefore, I am not listing the aforementioned commit as a Fixes commit.

Backport hint: this patch will have a trivial conflict applying to
stable kernels <= v5.4.

Reported-by: syzbot+ec07f6f5ce62b858579f@syzkaller.appspotmail.com
Tested-by: syzbot+ec07f6f5ce62b858579f@syzkaller.appspotmail.com
Closes: https://lore.kernel.org/linux-unionfs/671fd40c.050a0220.4735a.024f.GAE@google.com/
Cc: stable@vger.kernel.org
Signed-off-by: Amir Goldstein <amir73il@gmail.com>