nfsd: map EBUSY for rename()
Commit 466e16f0920f3 ("nfsd: check for EBUSY from vfs_rmdir/vfs_unink.")
mapped EBUSY host error from rmdir/unlink operation to avoid unknown
error server warning for the case that file is a local mount point.

The exact same reason can cause EBUSY error on rename() so map this
error from rename() the same as it is mapped for unlink()/rmdir()
depending if the renamed source is a dir or non-dir.

Fixes: 466e16f0920f3 ("nfsd: check for EBUSY from vfs_rmdir/vfs_unink.")
Suggested-by: Chuck Lever <chuck.lever@oracle.com>
Link: https://lore.kernel.org/linux-nfs/50c4f76e-0d5b-41a7-921e-32c812bd92f3@oracle.com/
Cc: Trond Myklebust <trondmy@hammerspace.com>
Cc: NeilBrown <neilb@suse.de>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>