fuse: introduce FUSE_PASSTHROUGH_INO mode
This is a more strict variant of FUSE_PASSTHROUGH mode, in which the
backing file inode number must match the fuse inode number.

This mode will allow the kernel to instantiate fuse inodes by
passthrough lookup and passthrough readdirplus and notify about those
inodes to the server, using the backing file inode number as a unique
identified for fuse inodes across kernel and server.

This mode limits the possiblity to map multiple fuse inodes to the same
backing file, unless they are all hardlinks.

This mode is only supported on 64bit arch, where ino_t is u64.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>